+++ dashboard/apps/dashboard/models/__init__.py
@@ -1,0 +1,9 @@
+# -*- coding: utf-8 -*-
+from .area import Area
+from .cost import (
+    AditionalCostsMixin, BaseCost, PersonCost, RatesManager, Rate, Cost,
+    Saving, Budget)
+from .link import Link
+from .person import Person
+from .product import Product, ProductGroup, ProductStatus, ProductGroupStatus
+from .task import Task
