+++ setup.cfg
@@ -1,9 +1,9 @@
 [flake8]
-exclude = .git/,env/,venv/,node_modules/, dashboard/apps/dashboard/migrations/,dashboard/settings/,dashboard/__init__.py
+exclude = .git/,env/,venv/,node_modules/, dashboard/apps/dashboard/migrations/,dashboard/settings/,dashboard/__init__.py,dashboard/apps/dashboard/models/__init__.py
 max-complexity = 10
 max-line-length = 120
 
-[pytest]
+[tool:pytest]
 DEBUG=True
 DJANGO_SETTINGS_MODULE=dashboard.settings.testing
 addopts = --ignore=venv --ignore=requirements
